TUCKER: Make _compressedSoundFlags uint16

+++ engines/tucker/tucker.h
@@ -387,7 +387,7 @@ public:
 private:
 
 	int _compressedSoundType;
-	int _compressedSoundFlags;
+	uint16 _compressedSoundFlags;
 	Common::File _fCompressedSound;
 };
